fix: missing styles
-Component card header fonts fixed after refactoring
-The accordion component image wasn't displaying for approx. 300px, so
extra queries are now in place to make sure it displays
-Some of the doodles were off-screen at screen sizes less the 440px,
and with some refactoring, now they stay on-screen at those small sizes.
-The "definition" background text now more closely matches the Figma file
-There was a small window of pixels where the related components arrow
was not near the section heading. Adding the extra media query with new
breakpoint kept the arrow closer to the heading
-Adding the transparent border allows the user to put focus on the code
example blocks without it shifting slightly
-Fixed inconsistent breakpoint changes
-Margin & padding for banner for interior pages matches designs better
-Match homepage breakpoints
-New SVG's for the background "definition" text were implemented
-Convert percents used on some of the homepage doodles to hardcoded rems
-Add min-width & min-height to doodles
-Added linear gradient to container in Definition component, instead of
to the entire component. This allows for less repetitive queries.
-Because line-height is unitless, some line-heights that were added
during this PR have been fixed to appropriately show the relationship
between the line-height and the font-size
-Increase thickness of outlines to meet accessibility guidelines

FSA22V2-301