fix: Homepage link & banner adjustments
-Fix where the header is located on each page. It no longer gets
imported as a completely separate component via Layout. It needs to be
implemented with the appropriate hero components per page in order to
inherit the correct background image.
-Working with the designer, a new layout for the illustrations was
created. She requested solid backgrounds behind the 3 component
illustrations.
-Since the header component has been relocated, several tests had to be
either updated or created. In Layout, the header was no longer being
rendered by that component any longer, so that test was updated and
moved into a separate Header.test file. Additionally, the image tests
within DetailsBanner needed to get updated because of the additional
grid images now in place.
-By adding an extra function to check if a component's illustration is
"wide," we can more dynamically determine whether or not the banner
has 3 images or a single, wide image.
-Add tests to check for proper results from the isWideImage function, and
checks for how many images get displayed in the banner.

FSA22V2-304