fix: api key access issue
For some reason, setting the `base` variable was throwing an error on
the home page, but not the component pages. Moving the variable
declaration into the `getComponentDetails` and `getHomePageInfo` seems
to resolve the issue.

Changing the functions to named exports instead of having one name and
one default was done for consistency more than anything.

AC-19